Move mock Expression to Test Utilities class
