+++ src/Analyzer.JsonRuleEngine.UnitTests/AllOfExpressionTests.cs
@@ -10,29 +10,13 @@ using Microsoft.Azure.Templates.Analyzer.Types;
 using Microsoft.Azure.Templates.Analyzer.Utilities;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq;
+using static Microsoft.Azure.Templates.Analyzer.RuleEngines.JsonEngine.UnitTests.TestUtilities;
 
 namespace Microsoft.Azure.Templates.Analyzer.RuleEngines.JsonEngine.UnitTests
 {
     [TestClass]
     public class AllOfExpressionTests
     {
-        /// <summary>
-        /// A mock implementation of an <see cref="Expression"/> for testing internal methods.
-        /// </summary>
-        private class MockExpression : Expression
-        {
-            public Func<IJsonPathResolver, JsonRuleEvaluation> EvaluationCallback { get; set; }
-
-            public MockExpression(ExpressionCommonProperties commonProperties)
-                : base(commonProperties)
-            { }
-
-            public override JsonRuleEvaluation Evaluate(IJsonPathResolver jsonScope)
-            {
-                return base.EvaluateInternal(jsonScope, EvaluationCallback);
-            }
-        }
-
         [DataTestMethod]
         [DataRow(true, true, DisplayName = "AllOf evaluates to true (true && true)")]
         [DataRow(true, false, DisplayName = "AllOf evaluates to false (true && false)")]
