+++ src/Analyzer.JsonRuleEngine.UnitTests/ExpressionTests.cs
@@ -1,34 +1,17 @@
 ﻿// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-using System;
 using Microsoft.Azure.Templates.Analyzer.RuleEngines.JsonEngine.Expressions;
 using Microsoft.Azure.Templates.Analyzer.Types;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq;
+using static Microsoft.Azure.Templates.Analyzer.RuleEngines.JsonEngine.UnitTests.TestUtilities;
 
 namespace Microsoft.Azure.Templates.Analyzer.RuleEngines.JsonEngine.UnitTests
 {
     [TestClass]
     public class ExpressionTests
     {
-        /// <summary>
-        /// A mock implementation of an <see cref="Expression"/> for testing internal methods.
-        /// </summary>
-        private class MockExpression : Expression
-        {
-            public Func<IJsonPathResolver, JsonRuleEvaluation> EvaluationCallback { get; set; }
-
-            public MockExpression(ExpressionCommonProperties commonProperties)
-                : base(commonProperties)
-            { }
-
-            public override JsonRuleEvaluation Evaluate(IJsonPathResolver jsonScope)
-            {
-                return base.EvaluateInternal(jsonScope, EvaluationCallback);
-            }
-        }
-
         [TestMethod]
         public void Evaluate_WhereConditionFalse_PathNotEvaluated()
         {
