+++ src/Analyzer.JsonRuleEngine.UnitTests/TestUtilities.cs
@@ -1,6 +1,9 @@
 ﻿// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+using System;
+using Microsoft.Azure.Templates.Analyzer.RuleEngines.JsonEngine.Expressions;
+using Microsoft.Azure.Templates.Analyzer.Types;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 
@@ -18,6 +21,23 @@ namespace Microsoft.Azure.Templates.Analyzer.RuleEngines.JsonEngine.UnitTests
         {
             var jsonValue = JsonConvert.SerializeObject(value);
             return JToken.Parse($"{{\"Key\": {jsonValue} }}")["Key"];
+        }
+
+        /// <summary>
+        /// A mock implementation of an <see cref="Expression"/> for testing internal methods.
+        /// </summary>
+        internal class MockExpression : Expression
+        {
+            public Func<IJsonPathResolver, JsonRuleEvaluation> EvaluationCallback { get; set; }
+
+            public MockExpression(ExpressionCommonProperties commonProperties)
+                : base(commonProperties)
+            { }
+
+            public override JsonRuleEvaluation Evaluate(IJsonPathResolver jsonScope)
+            {
+                return base.EvaluateInternal(jsonScope, EvaluationCallback);
+            }
         }
     }
 }
